manifest: remove SumAggregator
We also move the test-only NumFilesAnnotator to the test package. This
move triggered some false positives for `gcassert` which the rest of
the commit addresses (by including the `.` package).
